+++ src/app/App.js
@@ -7,7 +7,7 @@ import { useTranslation } from "react-i18next"
 import React, { useRef, useEffect } from "react"
 import { useSelector, useDispatch } from "react-redux"
 import { Banner, NotificationArea, MainContent, Footer, ScrollToTop } from "../components"
-import LoadingOverlay from "fork-victorvhn-react-loading-overlay"
+import LoadingOverlay from "react-loading-overlay"
 import {
     setAuth0IdToken,
     setAuth0AccessToken,